+++ src/test/java/io/github/astrapi69/throwable/ExceptionExtensionsTest.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.throwable;
+package io.github.astrapi69.throwable;
 
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;