+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 ## Change log
 ----------------------
 
-Version 1.5-SNAPSHOT
+Version 1.5
 -------------
 
 CHANGED:
 
+- update of gradle version to 6.7
+- update of test dependency testng to new version 7.3.0
 - moved method ThrowableExtensions#toRuntimeExceptionIfNeeded to appropriate RuntimeExceptionDecorator class with a new name decorate
+- changed to new package io.github.astrapi69
+- removed deprecated methods
 
 Version 1.4
 -------------